# ops/api/dappswagger.yaml
openapi: 3.0.0
servers:
  - url: 'http://52.15.91.107:3002/api/v1'
info:
  description: >-
    除注册与登陆两个API外，其他API的访问均需要在request header加上JWT token，否则返回`401 Unauthorized`。

    登陆成功后Server会返回JWT token。

    JWT token的格式为：
      - Value `Authorization: Bearer <token>` from request header

    **注意Bearer和token之前有一个空格**


    当使用swagger-ui访问API时，需要点右下方绿色`Authorize`按钮，输入`Bearer <token>`
    
    
    - 铸币账户：新建的token全部在这个账户里
    
    - gas账户：交易产生的所有gas都存到这个账户里
    
    - ram账户：所有ram都从这个账户里支出
    
    
    
    Sample Dapp code: https://github.com/XDMu/SampleDapp
    
    

    # change log __v0.0.2__
    
    本次版本主要更新如下：
    
    - Earth部署成功后需要初始化，初始化之前需要创建两个Earth账户，作为GZH的铸币账户和gas账户
    
    - 只有Admin才能创建token，创建token时需要指定3个账户  
    
    - 交易有gas和ram
    
    - admin可以改ram
    
    - 铸币账户可以改gas
    
    - Dapp部署之后需要到Earth注册一个合约账户
    
    - Dapp可以通过Earth的API接口进行转账操作

  version: 0.0.2
  title: Dapp API 文档
...
paths:
...
  /bureau/init:
    post:
      tags:
        - Bureau
      summary: 初始化智能合约账户
      description: 使用未来局需要初始化合约账户
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitContractAccountResponse'
        '401':
          description: UnAuthorized
        '400':
          description: Validation exception
        '500':
          description: Error
  /bureau/create:
    post:
      tags:
        - Bureau
      summary: 在区块链中创建未来局
      description: >-
        区块链的管理员用户可以授权给某个区块链用户**创建未来局**的权限（通过PUT
        /user/{username}），只有具备**创建未来局**权限的用户才能调用此API创建新的Token。
      operationId: addBureau
      security:
        - Bearer: []
      responses:
        '201':
          description: Success
        '401':
          description: UnAuthorized
        '405':
          description: Validation exception
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BureauBasic'
        description: Bureau object that needs to be added to the blockchain network
        required: true
  /bureau/query:
    post:
      summary: 查询未来局
      tags:
        - Bureau
      description: 查询未来局
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bureau'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryBureauRequest'
        description: 当前用户查询未来局
        required: true
  /bureau/getAll:
    get:
      summary: 获取所有未来局的名称
      tags:
        - Bureau
      description: 获取所有未来局的名称
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllBureauResponse'
  /bureau/betTransfer:
    post:
      tags:
        - Bureau
      summary: 未来局下注
      description: ''
      operationId: betBureau
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BetBureauResponse'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BetBureauRequest'
        required: true
  /bureau/settle:
    post:
      tags:
        - Bureau
      summary: 未来局结果判定
      description: 未来局结果判定
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgeBureauResponse'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JudgeBureauRequest'
        required: true
  /bureau/queryOption:
    post:
      tags:
        - Bureau
      summary: 查询某个用户某个未来局各个选项下注
      description: 查询某个用户某个未来局各个选项下注
      responses:
        '200':
          description: 成功
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - futureBureauName
              properties:
                username:
                  type: string
                futureBureauName:
                  type: string
  /bureau/queryParticipate:
    post:
      tags:
        - Bureau
      summary: 查询某个用户参与的未来局
      description: 查询某个用户某个未来局各个选项下注
      responses:
        '200':
          description: 成功
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
  /bureau/test:
    post:
      tags:
        - Bureau
      summary: 测试专用接口
      description: 测试专用接口
      responses:
        '200':
          description: 成功
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
  # 用户管理
...
  /auth/login:
    post:
      summary: 使用用户名密码登陆，返回JWT token
      description: 登陆
      tags:
        - Auth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        description: 用户名密码
        required: true
...
components:
  securitySchemes:
    Bearer:
      type: 'apiKey'
      name: Authorization
      in: header
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: string
          description: ''
        payload:
          type: object
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: 用户名
        password:
          type: string
          description: 密码，目前版本密码格式不加validation
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - payload
          properties:
            payload:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: 登陆成功返回的JWT token
    BureauBasic:
      type: object
      properties:
        name:
          type: string
          description: 未来局的英文名称
        chineseName:
          type: string
          description: 未来局的中文名称
        content:
          type: string
          description: 未来局的内容
        endTime:
          type: string
          description: 未来局结束时间
        option1:
          type: string
          description: 未来局选项一
        option2:
          type: string
          description: 未来局选项二
        option3:
          type: string
          description: 未来局选项三
        option4:
          type: string
          description: 未来局选项四
        option5:
          type: string
          description: 未来局选项五
        judgePerson:
          type: string
          description: 未来局裁定人
    Bureau:
      description: 未来局的data schema
      allOf: 
        - $ref: '#/components/schemas/BureauBasic'
        - type: object
          properties:
            creator:
              type: string
              description: 未来局创建者
            createTime:
              type: string
              description: 未来局创建时间
            history:
              type: array
              items:
                $ref: '#/components/schemas/BureauHistory'
    BureauHistory:
      description: 用户竞猜记录
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        token:
          type: string
          description: 下注用的代币
        chooseOption:
          type: string
          description: 选择的选项
        amount:
          type: number
          description: 下注的数量
...
    QueryBureauRequest:
      type: object
      properties:
        futureBureauName:
          type: string
          description: 未来局名称
    BetBureauRequest:
      type: object
      properties:
        futureBureauName:
          type: string
          description: 未来局的名称
        chooseOption:
          type: string
          description: 未来局选项
        amount:
          type: number
          description: 下注数量
        password:
          type: string
          format: password
          description: 输入密码以验证本次交易
    BetBureauResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
    InitContractAccountResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
    JudgeBureauRequest:
      type: object
      properties:
        futureBureauName:
          type: string
          description: 未来局的名称
        result:
          type: string
          description: 未来局的结果
    JudgeBureauResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
    AllBureauResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            payload:
              properties:
                bureaus:
                  type: array
                  items:
                    $ref: '#/components/schemas/MinusBureauInfo'
    MinusBureauInfo:
      type: object
      properties:
        name: 
          type: string
          description: 未来局名称
        creator:
          type: string
          description: 未来局创建者
        judgePerson:
          type: string
          description: 未来局结果判定人
        endTime:
          type: string
          description: 未来局结束时间